requirements.txt
Pillow>=9.5.0 